+++ sway/desktop/xdg_shell_v6.c
@@ -3,10 +3,10 @@
 #include <stdlib.h>
 #include <wayland-server.h>
 #include <wlr/types/wlr_xdg_shell_v6.h>
-#include "sway/container.h"
-#include "sway/layout.h"
+#include "sway/tree/container.h"
+#include "sway/tree/layout.h"
 #include "sway/server.h"
-#include "sway/view.h"
+#include "sway/tree/view.h"
 #include "sway/input/seat.h"
 #include "sway/input/input-manager.h"
 #include "log.h"
