+++ sway/input/cursor.c
@@ -7,7 +7,7 @@
 #include <wlr/types/wlr_cursor.h>
 #include <wlr/types/wlr_xcursor_manager.h>
 #include "sway/input/cursor.h"
-#include "sway/view.h"
+#include "sway/tree/view.h"
 #include "list.h"
 #include "log.h"
