+++ sway/main.c
@@ -18,7 +18,7 @@
 #include <wlr/util/log.h>
 #include "sway/config.h"
 #include "sway/server.h"
-#include "sway/layout.h"
+#include "sway/tree/layout.h"
 #include "sway/ipc-server.h"
 #include "ipc-client.h"
 #include "readline.h"
